+++ src/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'Netopia Payments for Vanilo',
-    'version' => '3.1-dev',
+    'version' => '3.1.0',
 ];